+++ src/app/main-page/components/MainLayoutClient/styles.ts
@@ -1,0 +1,6 @@
+export const mainLayoutStyles = {
+  container: 'min-h-screen flex',
+  alertContainer: 'fixed top-6 right-6 z-50 w-[400px]',
+  content: 'flex-grow',
+  main: 'p-6 '
+} as const;